Miscellaneous refactoring
- Fix Message and EngineData assignment operator (support self assignment).

- Simplify Component construction.

- Remove duplicated method to get current time. Move it to clara::msg::util
  and then add it to the clara::util scope.

- Fix usage of thread-unsafe C functions.

- Fix some bugs in json_utils.

- Remove unused method for address validation.

- Replace magic numbers with constants.

- Improve code of zhelper functions and try to optimize replyTo ID string
  creation.

- Restore null metadata check in Message constructor, that was moved to the body
  because of a bug in GCC.

- And other small changes and linter fixes.

+++ include/clara/msg/proto/data.hpp
@@ -194,7 +194,7 @@ inline T parse_data(const Data& data)
 inline std::vector<std::uint8_t> serialize_data(const Data& data)
 {
     auto buffer = std::vector<std::uint8_t>(data.ByteSizeLong());
-    data.SerializeToArray(buffer.data(), buffer.size());
+    data.SerializeToArray(buffer.data(), static_cast<int>(buffer.size()));
     return buffer;
 }
 
@@ -210,7 +210,7 @@ inline std::vector<std::uint8_t> serialize_data(const Data& data)
 inline Data parse_data(const std::vector<std::uint8_t>& buffer)
 {
     auto xdata = Data{};
-    xdata.ParseFromArray(buffer.data(), buffer.size());
+    xdata.ParseFromArray(buffer.data(), static_cast<int>(buffer.size()));
     return xdata;
 }
